added grid to main UI so we can stop absolute positioning the side bar. cleaned up buttons and list.
