+++ src/Wallet.js
@@ -1,27 +1,7 @@
 import React, { useState } from 'react'
 import Login from './Login'
 import useInterval from './useInterval'
-import { makeStyles } from '@material-ui/core/styles';
 import {Card, Typography, Grid} from '@material-ui/core'
-//import {SendIcon} from '@material-ui/core'
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 1,
-  },
-  paper: {
-    padding: theme.spacing(2),
-    textAlign: 'center',
-    color: theme.palette.text.secondary,
-  },
-  large:{
-    height: '100px',
-    width: "100px", 
-    margin: "auto"
-  }
-}));
-
-
 function Wallet({ computer }) {
   const [balance, setBalance] = useState(0)
 
@@ -31,7 +11,6 @@ function Wallet({ computer }) {
     }
     getBalance()
   }, 3000)
-  //const classes = useStyles();
   return (
     <div>
     <Card className='flex'>
